backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-adk
psycopg2-binary
